Fix/number exception (#66)
* Fixed parsing of number parametters recieved from react-native

* Removed useless method Temp and use react-native one instead

* Used getDouble instead of getInt because it can cause cast exception

* Revert "Removed useless method Temp and use react-native one instead"

This reverts commit 0dee580cd279f9ae78d29cd6794c80aba4f19b66.

* bring back the old Temp class

+++ android/src/main/java/guichaguri/trackplayer/metadata/Metadata.java
@@ -75,8 +75,8 @@ public class Metadata {
         notification.updateOptions(data);
 
         // Load the options
-        ratingType = data.getInt("ratingType", ratingType);
-        maxArtworkSize = data.getInt("maxArtworkSize", maxArtworkSize);
+        ratingType = (int)data.getDouble("ratingType", ratingType);
+        maxArtworkSize = (int)data.getDouble("maxArtworkSize", maxArtworkSize);
 
         // Update the rating type
         session.setRatingType(ratingType);
